% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fxn_call_api}
\alias{fxn_call_api}
\title{Call the xkcd API}
\usage{
fxn_call_api(x)
}
\arguments{
\item{x}{Optional: an integer, corresponding to the xkcd comic number. If
nothing is passed, the latest comic is returned.}
}
\value{
An object of class \code{gt_tbl} from {gt} package.
}
\description{
This is a helper function to call the xkcd API
}
\examples{
\dontrun{
fxn_call_api(1)
}
}
