% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fxn_json_to_tibble}
\alias{fxn_json_to_tibble}
\title{Convert xkcd JSON to a tibble}
\usage{
fxn_json_to_tibble(x)
}
\arguments{
\item{x}{The returned response JSON object.}
}
\value{
tibble
}
\description{
This function converts the returned JSON file from xkcd into a tibble.
}
\examples{
\dontrun{
resp <- httr::GET(endpoint)
fxn_json_to_tibble(resp)
}
}
