% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fxn_latest_comic_number}
\alias{fxn_latest_comic_number}
\title{Get the number of the latest xkcd comic}
\usage{
fxn_latest_comic_number()
}
\value{
string
}
\description{
This function gets the comic number for the latest xkcd
}
\examples{
\dontrun{
fxn_latest_comic_number()
}
}
