% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fxn_make_table}
\alias{fxn_make_table}
\title{Make xkcd table}
\usage{
fxn_make_table(x)
}
\arguments{
\item{x}{The dataframe obtained.}
}
\value{
An object of class \code{gt_tbl} from {gt} package.
}
\description{
This is the main function to create the xkcd comic panel.
}
\examples{
\dontrun{
fxn_call_api(1) |>
  fxn_json_to_tibble() |>
  fxn_make_table()
}
}
