% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xkcd.R
\name{get_xkcd}
\alias{get_xkcd}
\title{Get xkcd comic}
\usage{
get_xkcd(x = NA, ...)
}
\arguments{
\item{x}{Optional: an integer, corresponding to the xkcd comic number. If
nothing is passed, the latest comic is returned.
Defaults to \code{400}.}

\item{...}{Unused arguments, reserved for future expansion.}
}
\value{
An object of class \code{gt_tbl} from {gt} package.
}
\description{
This is the main function to call for retrieving xkcd comics.
}
\examples{
\dontrun{
get_xkcd(1)
}
}
